Create MainProject_update.html
For this update the background still has to be resized to make the page look really nice. I do have a help page created in one of my files and I plan to connect that page to a link called "About" on the main program's page. That won't be hard, I just want to get the background image resized. Lastly, I am trying to use wrappers to display the data in, but I'm still not sure yet if that will work. If not I will go a different route, if so then it's just a matter of formatting