Create MainPage.html
This is the edited Main page for the site. It has an about link that explain the site's purpose, but I switched some things around. I don't know what the output is gonna look like so I simply have an upload box than calculate the file. Once calculate is hit we can transfer to another page that displays the information. I wanted to have the Gantt chart appear on this main page after calculate was hit and then have a drop down menu with categories of information that the user could be taken to, but depending on how the rest of th project goes we will see about that. I think it would be cool.